build: add #[deny(unused_crate_dependencies)] to all libraries
This [attribute](https://doc.rust-lang.org/reference/attributes.html)
detects unused dependencies. It can give false positives
(https://doc.rust-lang.org/rustc/lints/listing/allowed-by-default.html#unused-crate-dependencies),
so I restructured some crates to avoid this (moving `main` into `ipso-cli`,
moving `benchmarks` out of `ipso-cli`).
