+++ ipso-eval/Cargo.toml
@@ -8,8 +8,6 @@ edition = "2018"
 [dependencies]
 fnv = "1.0.7"
 ipso-core = { path = "../ipso-core" }
-ipso-import = { path = "../ipso-import" }
 ipso-rope = { path = "../ipso-rope" }
 ipso-syntax = { path = "../ipso-syntax" }
 paste = "1.0.5"
-typed-arena = "2.0.1"
