+++ ipso-repl/Cargo.toml
@@ -11,4 +11,3 @@ ipso-eval = { path = "../ipso-eval" }
 ipso-import = { path = "../ipso-import" }
 ipso-syntax = { path = "../ipso-syntax" }
 ipso-typecheck = { path = "../ipso-typecheck" }
-typed-arena = "2.0.1"
