+++ ipso-typecheck/Cargo.toml
@@ -11,7 +11,6 @@ ipso-core = { path = "../ipso-core" }
 ipso-diagnostic = { path = "../ipso-diagnostic" }
 ipso-rope = { path = "../ipso-rope" }
 ipso-syntax = { path = "../ipso-syntax" }
-ipso-util = { path = "../ipso-util" }
 
 [dev-dependencies]
 pretty_assertions = "1.3.0"
