+++ ipso-util/src/lib.rs
@@ -1,4 +1,5 @@
+#![deny(unused_crate_dependencies)]
+
 pub mod fnv_hash_multi_map;
 pub mod hash_multi_set;
 pub mod iter;
-pub mod void;
