--- ipso-util/src/void.rs
@@ -1,7 +1,0 @@
-pub enum Void {}
-
-impl Void {
-    pub fn absurd<A>(&self) -> A {
-        panic!("void")
-    }
-}
